Added cube function and used it inside Blocks.cpp (more info bellow)
This cube function is based on what Taras put inside the Blocks.cpp file.
It draws a 1x1x1 cube centered at the origin that has normals and texture
coordinates built in.

To use:
*Include Cube.h in your file
*Wherever you would want to draw a cube call pj::cube()

Warning: The function includes glBegin(GL_QUADS) at the beginning and
glEnd() at the end so it may disrupt those functions if you put it between
those.

Also it begins with glPushMatrix() and ends with glPopMatrix() so there is
no need to put those around the function when you use it.